+++ CampaignService/src/main/java/com/example/CampaignService/controller/VisekratnaKampanjaController.java
@@ -4,6 +4,7 @@ import com.example.CampaignService.TokenUtils;
 import com.example.CampaignService.model.*;
 import com.example.CampaignService.repository.AgentRepository;
 import com.example.CampaignService.service.VisekratnaKampanjaService;
+import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -27,11 +28,11 @@ public class VisekratnaKampanjaController {
 
     @CrossOrigin(origins = "http://localhost:4200")
     @PostMapping("/create")
-    public VisekratnaKampanja create(@RequestBody VisekratnaDTO visekratnaDTO, @RequestHeader(value="Authorization") String token) throws Exception {
+    public ResponseEntity<VisekratnaKampanja> create(@RequestBody VisekratnaDTO visekratnaDTO, @RequestHeader(value="Authorization") String token) throws Exception {
         String role = tokenUtils.getRoleFromToken(token);
         if (role.equals("agent")) {
             Integer userId = tokenUtils.getIdFromToken(token);
-            return visekratnaKampanjaService.saveKampanja(visekratnaDTO, userId);
+            return new ResponseEntity<>(visekratnaKampanjaService.saveKampanja(visekratnaDTO, userId), HttpStatus.OK);
         } else {
             throw new Exception("Zabranjeno");
         }
@@ -39,21 +40,36 @@ public class VisekratnaKampanjaController {
 
     @CrossOrigin(origins = "http://localhost:4200")
     @GetMapping("/getAll")
-    public List<KampanjaReturnDTO> getAll(@RequestHeader(value="Authorization") String token) {
+    public ResponseEntity<List<KampanjaReturnDTO>> getAll(@RequestHeader(value="Authorization") String token) throws Exception {
         Integer userId = tokenUtils.getIdFromToken(token);
-        return visekratnaKampanjaService.getAllByAgent(agentRepository.findOneById(userId));
+        String role = tokenUtils.getRoleFromToken(token);
+        if (role.equals("agent")) {
+            return new ResponseEntity<>(visekratnaKampanjaService.getAllByAgent(agentRepository.findOneById(userId)), HttpStatus.OK);
+        } else {
+            throw new Exception("Zabranjeno");
+        }
     }
 
     @CrossOrigin(origins = "http://localhost:4200")
     @PostMapping("/change")
-    public ResponseEntity<VisekratnaKampanja> changeKampanja(@RequestBody VisekratnaDTO visekratnaDTO) {
-        return new ResponseEntity<>(visekratnaKampanjaService.changeKampanja(visekratnaDTO), HttpStatus.OK);
+    public ResponseEntity<VisekratnaKampanja> changeKampanja(@RequestBody VisekratnaDTO visekratnaDTO, @RequestHeader(value="Authorization") String token) throws Exception {
+        String role = tokenUtils.getRoleFromToken(token);
+        if (role.equals("agent")) {
+            return new ResponseEntity<>(visekratnaKampanjaService.changeKampanja(visekratnaDTO), HttpStatus.OK);
+        } else {
+            throw new Exception("Zabranjeno");
+        }
     }
 
     @CrossOrigin(origins = "http://localhost:4200")
     @PostMapping("/delete")
-    public void deleteKampanja(@RequestBody IdDTO idDTO) {
-        visekratnaKampanjaService.deleteKampanja(idDTO);
+    public void deleteKampanja(@RequestBody IdDTO idDTO, @RequestHeader(value="Authorization") String token) throws Exception {
+        String role = tokenUtils.getRoleFromToken(token);
+        if (role.equals("agent")) {
+            visekratnaKampanjaService.deleteKampanja(idDTO);
+        } else {
+            throw new Exception("Zabranjeno");
+        }
     }
 
 }
